added shoping cart

+++ database/migrations/2018_03_05_112537_create_dishes_table.php
@@ -18,7 +18,7 @@ class CreateDishesTable extends Migration
             $table->string('title');
             $table->text('description');
             $table->string('image_url');
-            $table->double('price', 7, 2);
+            $table->float('price', 8, 2);
             $table->timestamps();
         });
     }
